New version 1.3.0
Proved to work with MediaWiki 1.19, using ResourceLoader

Rewrote the way _legacyKeyDown was being used. It was sooooo dirty...

It was done getting the original keydown event from the jQuery internal
object, that doesn't seem to work now (something changed), so I prefer to
just copy the contents of that funcion into this script and update when
needed, since the old way was going to break sooner or later.